refactor: introduce UserSettings type and update constants
- Added UserSettings interface in a new types.ts file to define user settings structure.
- Introduced DEFAULT_SETTINGS constant in constants.ts for default user settings.
- Updated imports in SettingsDialog and canvas files to reference the new constants and types, enhancing code organization and type safety.
