+++ src/frontend/src/lib/canvas.ts
@@ -1,4 +1,4 @@
-import { DEFAULT_SETTINGS } from '../types/settings';
+import { DEFAULT_SETTINGS } from '../constants';
 
 /**
  *
